config popular ui.
