+++ MyApp/View/Popular/Child/ChildPopularViewController.swift
@@ -1,0 +1,45 @@
+//
+//  ChildPopularViewController.swift
+//  MyApp
+//
+//  Created by ANH NGUYỄN on 2/15/20.
+//  Copyright © 2020 Asian Tech Co., Ltd. All rights reserved.
+//
+
+import UIKit
+
+final class ChildPopularViewController: ViewController {
+
+    @IBOutlet weak var tableView: UITableView!
+    
+    var viewModel = ChildPopularViewModel()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+    
+    override func setupUI() {
+        tableView.register(name: CellIdentifier.homeCell.rawValue)
+        tableView.dataSource = self
+        tableView.delegate = self
+    }
+}
+extension ChildPopularViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 10
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: CellIdentifier.homeCell.rawValue, for: indexPath) as? HomeCell else {
+            return UITableViewCell()
+        }
+        cell.viewModel = viewModel.viewModelForCell(at: indexPath)
+        return cell
+    }
+}
+
+extension ChildPopularViewController: UITableViewDelegate {
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        UITableView.automaticDimension
+    }
+}
