+++ MyApp/View/Popular/PopularViewController.swift
@@ -9,9 +9,11 @@
 import UIKit
 
 final class PopularViewController: ViewController {
-
-    @IBOutlet weak var tableView: UITableView!
-
+    
+    @IBOutlet weak var segmentControl: UISegmentedControl!
+    @IBOutlet weak var contentView: UIView!
+    let child = ChildPopularViewController()
+    
     var viewModel = PopularViewModel()
 
     override func viewDidLoad() {
@@ -20,25 +22,8 @@ final class PopularViewController: ViewController {
 
     override func setupUI() {
         super.setupUI()
-        tableView.register(name: CellIdentifier.homeCell.rawValue)
-        tableView.delegate = self
-        tableView.dataSource = self
-    }
-}
-extension PopularViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return viewModel.videos.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: CellIdentifier.homeCell.rawValue, for: indexPath) as? HomeCell else { return UITableViewCell() }
-        cell.viewModel = viewModel.viewModelForCell(at: indexPath)
-        return cell
-    }
-}
-
-extension PopularViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return UITableView.automaticDimension
+        addChild(child)
+        view.addSubview(child.view)
+        child.didMove(toParent: self)
     }
 }
